requirements.txt
# N8N Workflows API Dependencies
# Core API Framework - Stable versions compatible with Python 3.9-3.12
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP & Networking
httpx==0.26.0
requests==2.31.0

# Monitoring & Performance
psutil==5.9.8

# Email validation
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# N8N Workflows API Dependencies
# Core API Framework - Stable versions compatible with Python 3.9-3.12
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP & Networking
httpx==0.26.0
requests==2.31.0

# Monitoring & Performance
psutil==5.9.8

# Email validation
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# Agent Capabilities
pyautogen
openai
supabase
gitpython
beautifulsoup4
requests
pandas
# activepieces-python-client>=0.1.0  # Removed due to install error, using requests instead
